+++ adapters/LiftoffMonetize/LiftoffMonetizeAdapter/Bidding/GADMediationVungleInterstitial.m
@@ -143,7 +143,7 @@
 }
 
 - (void)interstitialAdWillLeaveApplication:(nonnull VungleInterstitial *)interstitial {
-  [_delegate willBackgroundApplication];
+  // Google Mobile Ads SDK doesn't have a matching event.
 }
 
 #pragma mark - GADMAdapterVungleDelegate