+++ adapters/LiftoffMonetize/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Liftoff Monetize iOS Mediation Adapter Changelog
 
+#### Next version
+- Removed the deprecated willBackgroundApplication delegate methods from the banner and interstitial.
+
 #### [Version 7.1.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/liftoffmonetize/LiftoffMonetizeAdapter-7.1.0.0.zip)
 - Verified compatibility with Vungle SDK 7.1.0.